+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2025-05-30
+
+### Changed
+
+- Applied WP naming conventions to files, functions, css classes. 
 
 ## [1.1.1] - 2025-05-23
 
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Date of Baptism metadata for all members - DoBap
 - Added encoding function to ensure special characters in phone numages and email address display correctly. 
 - Added dynamic column heading for data export permitting up to specified number of family members. 
-- Add maxFamilyMembers constant in class-FamilyDirectoryApp.php to define max size of members db entry. 
+- Add maxFamilyMembers constant in class-COTA_Family_Directory_App.php to define max size of members db entry. 
 
 ### Changed